Rename: 주석 정리 및 파일 위치 이동
