+++ src/main/java/nbc/chillguys/nebulazone/infra/websocket/AuthenticationChannelInterceptor.java
@@ -29,36 +29,41 @@ public class AuthenticationChannelInterceptor implements ChannelInterceptor {
 	private final ChatRoomRepository chatRoomRepository;
 	private final ChatRoomUserRepository chatRoomUserRepository;
 
+	/**
+	 * WebSocket STOMP 인바운드 메시지 인증 및 권한 검증 인터셉터 <br>
+	 * <p>
+	 * - CONNECT: JWT 토큰을 검증하고, Principal 및 세션-유저 매핑을 수행<br>
+	 * - SUBSCRIBE: 채팅방 존재 여부와 참여자 권한을 검증하고, 세션-채팅방 매핑을 수행
+	 * </p>
+	 *
+	 * @author 박형우
+	 */
 	@Override
 	public Message<?> preSend(Message<?> message, MessageChannel channel) {
 		StompHeaderAccessor accessor = StompHeaderAccessor.wrap(message);
 
-		// STOMP CONNECT
+		// CONNECT: 최초 연결 시 JWT 인증 처리
 		if (StompCommand.CONNECT.equals(accessor.getCommand())) {
 
-			// 헤더에서 토큰 꺼내기
 			String token = accessor.getFirstNativeHeader("Authorization");
 
 			if (token == null) {
 				throw new IllegalArgumentException("No token provided");
 			}
 
-			// "Bearer " 제거
 			if (token.startsWith("Bearer ")) {
 				token = token.substring("Bearer ".length());
 			}
 
 			try {
-				// jWT 토큰 파싱 및 검증
 				AuthUser authUserFromToken = jwtUtil.getAuthUserFromToken(token);
 
-				// accessor.setUser에는 Principal 타입이 필요하기 때문에 UsernamePasswordAuthenticationToken으로 감싸기
 				Principal principal = new UsernamePasswordAuthenticationToken(String.valueOf(authUserFromToken.getId()),
 					null, authUserFromToken.getAuthorities());
 
 				accessor.setUser(principal);
 
-				// CONNECT 단계 에서는 유저 정보만 매핑
+				// 세션과 유저 매핑 (메모리 or Redis 등)
 				SessionUtil.registerUser(accessor.getSessionId(), authUserFromToken);
 
 			} catch (Exception e) {
@@ -68,10 +73,9 @@ public class AuthenticationChannelInterceptor implements ChannelInterceptor {
 
 		}
 
-		// STOMP SUBSCRIBE
+		// SUBSCRIBE: 특정 채팅방 구독 시 권한 체크
 		if (StompCommand.SUBSCRIBE.equals(accessor.getCommand())) {
 
-			// 구독할 경로 추출
 			String destination = accessor.getDestination();
 
 			if (destination != null && destination.startsWith("/topic/chat/")) {
@@ -80,19 +84,19 @@ public class AuthenticationChannelInterceptor implements ChannelInterceptor {
 					Long roomId = Long.valueOf(roomIdStr);
 					AuthUser authUser = SessionUtil.getUserIdBySessionId(accessor.getSessionId());
 
-					// 채팅방 존재 여부 확인
+					// 채팅방 존재 확인
 					boolean existsChatRoomById = chatRoomRepository.existsChatRoomById(roomId);
 					if (!existsChatRoomById) {
 						throw new ChatException(ChatErrorCode.CHAT_ROOM_NOT_FOUND);
 					}
 
-					// 채팅방에 참여중인 유저인지 확인
+					// 채팅방 참가자 여부 확인
 					boolean isParticipant = chatRoomUserRepository.existsByIdChatRoomIdAndIdUserId(roomId,
 						authUser.getId());
 					if (!isParticipant) {
 						throw new ChatException(ChatErrorCode.CHAT_ROOM_ACCESS_DENIED);
 					}
-
+					// 세션과 채팅방 매핑
 					SessionUtil.registerRoom(accessor.getSessionId(), roomId);
 				} catch (NumberFormatException e) {
 					log.warn("방번호 추출 실패: {}", roomIdStr);
